+++ src/ChangelogsPlugin.php
@@ -12,24 +12,21 @@
 namespace Pyrech\ComposerChangelogs;
 
 use Composer\Composer;
-use Composer\DependencyResolver\Operation\UpdateOperation;
 use Composer\EventDispatcher\EventSubscriberInterface;
+use Composer\Installer\PackageEvent;
+use Composer\Installer\PackageEvents;
 use Composer\IO\IOInterface;
-use Composer\Package\PackageInterface;
 use Composer\Plugin\PluginInterface;
 use Composer\Script\Event;
-use Composer\Script\PackageEvent;
 use Composer\Script\ScriptEvents;
-use Pyrech\ComposerChangelogs\UrlGenerator\BitbucketUrlGenerator;
-use Pyrech\ComposerChangelogs\UrlGenerator\GithubUrlGenerator;
 
 class ChangelogsPlugin implements PluginInterface, EventSubscriberInterface
 {
     /** @var IOInterface */
     private $io;
 
-    /** @var Changelogs */
-    private $changelogs;
+    /** @var Outputter */
+    private $outputter;
 
     /**
      * {@inheritdoc}
@@ -37,10 +34,7 @@ class ChangelogsPlugin implements PluginInterface, EventSubscriberInterface
     public function activate(Composer $composer, IOInterface $io)
     {
         $this->io = $io;
-        $this->changelogs = new Changelogs([
-            new GithubUrlGenerator(),
-            new BitbucketUrlGenerator(),
-        ]);
+        $this->outputter = Factory::createOutputter();
     }
 
     /**
@@ -48,42 +42,30 @@ class ChangelogsPlugin implements PluginInterface, EventSubscriberInterface
      */
     public static function getSubscribedEvents()
     {
-        return array(
-            ScriptEvents::POST_PACKAGE_UPDATE => array(
-                array('postPackageUpdate'),
-            ),
-            ScriptEvents::POST_UPDATE_CMD => array(
-                array('postUpdate'),
-            ),
-        );
+        return [
+            PackageEvents::POST_PACKAGE_UPDATE => [
+                ['postPackageOperation'],
+            ],
+            PackageEvents::POST_PACKAGE_INSTALL => [
+                ['postPackageOperation'],
+            ],
+            PackageEvents::POST_PACKAGE_UNINSTALL => [
+                ['postPackageOperation'],
+            ],
+            ScriptEvents::POST_UPDATE_CMD => [
+                ['postUpdate'],
+            ],
+        ];
     }
 
     /**
      * @param PackageEvent $event
      */
-    public function postPackageUpdate(PackageEvent $event)
+    public function postPackageOperation(PackageEvent $event)
     {
         $operation = $event->getOperation();
 
-        if (!$operation instanceof UpdateOperation) {
-            return;
-        }
-
-        /** @var PackageInterface $initialPackage */
-        $initialPackage = $operation->getInitialPackage();
-        /** @var PackageInterface $targetPackage */
-        $targetPackage = $operation->getTargetPackage();
-
-        if ($initialPackage->getName() !== $targetPackage->getName()) {
-            return;
-        }
-
-        $this->changelogs->addUpdate(new Update(
-            $initialPackage->getName(),
-            $initialPackage->getPrettyVersion(),
-            $targetPackage->getPrettyVersion(),
-            $targetPackage->getSourceUrl()
-        ));
+        $this->outputter->addOperation($operation);
     }
 
     /**
@@ -91,10 +73,6 @@ class ChangelogsPlugin implements PluginInterface, EventSubscriberInterface
      */
     public function postUpdate(Event $event)
     {
-        if ($this->changelogs->isEmpty()) {
-            return;
-        }
-
-        $this->io->write($this->changelogs->getOutput());
+        $this->io->write($this->outputter->getOutput());
     }
 }