+++ CHANGELOG.md
@@ -1,5 +1,11 @@
  # Changes between versions
  
+ ## Not yet released
+ 
+ * Add support for bitbucket repositories
+ * Add new line after plugin output
+ * Add support for install and uninstall operations
+ 
  ## 1.0 (2015-09-26)
  
  * Initial release